requirements.txt
django==4.2.1
ipython==8.13.2
djangorestframework==3.14.0
pillow==9.5.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
drf-spectacular==0.26.2
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
coverage==7.2.7
ruff==0.0.265
pre-commit==3.3.1
isort==5.10.1
black==23.1.0
